Permit archival of stacks
Archiving a stack hides it from the list of stacks in the index view (but not from API clients) and locks the stack for deployment.

This is useful in itself for uncluttering the UI while retaining data, but is also intended to form part of the future addition of an ability to auto-provision stacks based on open PRs; when a PR is closed, its stack would be archived, and resources allocated to deploying the stack (eg Kubernetes namespaces) would be destroyed at this point.