Add environment_url on successful deployment status
When creating a deployment status event for a successful deployment,
let's add the `environment_url` to the payload we send to GitHub. The
idea is that this gets us whatever awesome features with which the GitHub API
uses this value - automatic linking to the deployment environment in
the PR - but also we can then pass this through to systems consuming
the deployment events to provide additional functionality.

References
----------

- https://developer.github.com/changes/2016-04-06-deployment-and-deployment-status-enhancements/#link-to-a-live-deployment