Prevent GithubSyncJob from blowing up if commit already exists
My current thinking is that when we create a stack and associate the
PR we enqueue two jobs. The GithubSyncJob and the
RefreshPullRequestJob. These both run concurrently. In some cases, the
GithubSyncJob runs and attempts to find the list of commits for the
stack which haven't yet been created - this includes the PR's head ref:

https://github.com/powerhome/shipit-engine/blob/master/app/jobs/shipit/github_sync_job.rb#L32-L44

Then the RefreshPullRequestJob runs and creates the Shipit::Commit
object for the head ref of the branch:

https://github.com/powerhome/shipit-engine/blob/pr-review-stacks/app/models/shipit/pull_request.rb#L243

Then the GithubSyncJob gets scheduled and attempts to create the
Shipit::Commit object for the head ref and blows up:

https://github.com/powerhome/shipit-engine/blob/master/app/jobs/shipit/github_sync_job.rb#L29