Add ReviewStack into shipit engine
Makes ReviewStacks a 1st class citizen of shipit-engine. Additionally
serves as a place to start pulling the review stack functionality from
our host application down into shipit-engine.

We've not pulled functionality from the host application into this
body of work to minimize the risk to our existing users - IE let's
tinker with the engine while the car's running down the highway, but
let's leave the wheels alone for now. That is to say, there is much
more work coming.