Prevent updating labels for yet-to-be-created pull_requests
When labels are applied at the time of PullRequest creation, sometimes
the PullRequest Labeled event arrives to shipit before the PullRequest
is created. In this case, we want to drop the "labeld" message because
the pull request doesn't exist, AND we'll apply the labels via the
pull request creation anyway.