Remove no-op migrations
Over the course of implementing and evolving the ReviewStack feature
of shipit-engine we've evolved how we model the data. This cleans up
that data modeling history to present a clear statement of intended
changes to the existing schema of shipit-engine. This snapshot of the
end product seems more appropriate for inclusion in the upstream PR of
ReviewStacks rather than inclusion of every mis-step along the way.

This was tested by checking out the shopify/master branch to setup the
local database, then checking this branch out and running the
migrations. The fact that the schema.rb was only impacted by reverting
a change we didn't really want to make to the upstream schema gives me
confidence that this chain of migrations is functionally equivalent to
what was in place before.