Atomically create Review Stacks and their Pull Requests
Last evening we observed a Review Stack was created without a Pull
Request. It is our intent that EVERY Review Stack should have an
associated Pull Request. We suspect that the call to
`PullRequest#update!` failed. Since creating the ReviewStack and
PullRequest are flushed to the database independently, we ended up
with a ReviewStack that was missing its PullRequest. This change
attempts to make the creation of the Review Stack and its Pull Request
atomic - ensuring that any Review Stack that is created will have
captured its Pull Request.